test(user): add user tests, user validation; refactored User code
When input comes from an API endpoint, it will not be unmarshalled, but
rather it will be stored in an input object before being transitioned to
a User object.

Refactored UpdateUser and PatchUser.  The uuid will now be checked
against the User object being passed in rather than as its own
parameter.